\name{col.coolhot20}
\alias{col.coolhot20}
\alias{col.coolhot21}
\docType{data}
\title{ Color Palette }
\description{
  convenient color palettes for plotting indices to a map
}
\usage{
  data(col.coolhot20)
  data(col.coolhot21)
}
\format{
  character vector containing R color codes}
\details{
  col.coolhot20 is ten "cool" colors and 10 "hot" colors. col.coolhot21 is the same as col.coolhot20, but white added at col.coolhot21[11]. For a "oneway" color scheme, try rev(col.coolhot20[1:10]) or col.coolhot20[11:20].
}
\Author{
  Anthony L. Westerling
}
\references{
  in preparation
}
\keyword{color}
