\name{div.code}
\alias{div.code}
\docType{data}
\title{ Character Vector }
\description{
  a list of state division codes for each U.S. Climate Division of the form "CA01", where "CA" is the state code for California, and "01" denotes the first climate division in that state.
}
\usage{
div.code is used internally by functions in package climdiv.
}
\format{
  The format is:
 chr [1:359] "MN02" "WA04" "ND02" "ND01" "WA05" "MT03" "MT06" "WA03" ...
}
\details{
  "NA00" denotes polygons that represent empty space rather than a climate division.
}
\keyword{datasets}
