\name{western.ST}
\alias{western.ST}
\docType{data}
\title{ character vector}
\description{
  state abreviations for 11 western states in the contiguous United States.
}
\usage{data(western.ST)}
\format{
  The format is:
 chr [1:11] "CA" "NV" "AZ" "NM" "CO" "UT" "WY" "MT" "ID" "WA" "OR"
}
\details{
  useful for matching character strings to plot by state or clip to states.
}
\seealso{
  \code{\link{western.divs}}, \code{\link{western.states}}, \code{\link{div.code}}, \code{\link{climdiv.poly}}
}
\author{
  Anthony L. Westerling
}
\examples{
data(western.ST)
}
\keyword{datasets}
