\name{western.states}
\alias{western.states}
\docType{data}
\title{ character vector}
\description{
  state names for 11 western states in the contiguous United States.
}
\usage{data(western.states)}
\format{
  The format is:
 chr [1:11] "california" "nevada" "arizona" "new mexico" "colorado" ...
}
\details{
  useful for matching character strings to plot by state or clip to states.
}
\seealso{
  \code{\link{western.divs}}, \code{\link{western.ST}}, \code{\link{div.code}}, \code{\link{climdiv.poly}}
}
\author{
  Anthony L. Westerling
}
\examples{
data(western.states)
}
\keyword{datasets}
